update: function gerar excel
